Fix bug in move parsing
Caused due to greediness in matching alternatives that share prefixes. If
we have r"a|ab", then "a" will be matched first even if a "b" is present
later, as when matching "ab". Changing the ordering to put the longer
patterns first solves the issue, but feels like it's still easy to get
wrong.